feat: Chat Service  Rx/Tx Redis Transport (#34)
Adds Twitch-Redis Transport Service
Adds gha docker build 
Adds val build targets
Adds framework for mockserver
Adds disabled chat integration test

+++ scripts/package.sh
@@ -2,9 +2,13 @@
 # package.sh
 # Copyright (C) 2023
 # Squidpie
-
+if [[ -z ${GITVERSION_SEMVER} ]]; then
+    echo "Env Var GITVERSION_SEMVER not set"
+    exit -1
+fi
 RelName="strauss-${GITVERSION_SEMVER}"
-RelPath="packaging/${RelName}"
+RelRoot="packaging"
+RelPath="${RelRoot}/${RelName}"
 
 # Setup output dir
 if [ -d "${RelPath}" ]; then rm -Rf ${RelPath}; fi
@@ -21,5 +25,6 @@ cp .env ${RelPath}/
 # Copy deploy script
 cp scripts/deploy.sh ${RelPath}/
 
-tar czf ${RelName}.tar.gz -C packaging ${RelName}
-mv ${RelName}.tar.gz ${RelPath}/
+# Create Package & Cleanup
+tar czf ${RelPath}.tar.gz -C packaging ${RelName}
+rm -r ${RelPath}/
